fixup! feat(angular): handle angular components without selector
