+++ examples/angular-cli/src/stories/basics/component-without-selector/without-selector.stories.ts
@@ -1,43 +1,5 @@
-import {
-  Component,
-  Inject,
-  InjectionToken,
-  Injector,
-  Input,
-  OnInit,
-  Optional,
-  Type,
-} from '@angular/core';
-import { componentWrapperDecorator, moduleMetadata, Story, Meta } from '@storybook/angular';
-
-const WITHOUT_SELECTOR_DATA = new InjectionToken<{ color: string; name: string }>(
-  'WITHOUT_SELECTOR_DATA'
-);
-
-@Component({
-  template: `My name in color :
-    <div [style.color]="color">{{ name }}</div>
-    <ng-content></ng-content> <ng-content></ng-content>`,
-})
-class WithoutSelectorComponent {
-  color = '#1e88e5';
-
-  name = 'Joe Bar';
-
-  constructor(
-    @Inject(WITHOUT_SELECTOR_DATA)
-    @Optional()
-    data: {
-      color: string;
-      name: string;
-    } | null
-  ) {
-    if (data) {
-      this.color = data.color;
-      this.name = data.name;
-    }
-  }
-}
+import { moduleMetadata, Story, Meta } from '@storybook/angular';
+import { WithoutSelectorComponent, WITHOUT_SELECTOR_DATA } from './without-selector.component';
 
 export default {
   title: 'Basics / Component / without selector',
@@ -66,66 +28,3 @@ WithInjectionTokenAndArgs.argTypes = {
   color: { control: 'color' },
 };
 WithInjectionTokenAndArgs.args = { name: 'Dixie Normous', color: 'red' };
-
-// Advanced example with custom *ngComponentOutlet
-
-@Component({
-  selector: 'without-selector-wrapper',
-  template: `<ng-container
-    *ngComponentOutlet="componentOutlet; injector: componentInjector; content: componentContent"
-  ></ng-container>`,
-})
-class WithoutSelectorWrapperComponent implements OnInit {
-  @Input()
-  componentOutlet: Type<unknown>;
-
-  @Input()
-  name: string;
-
-  @Input()
-  color: string;
-
-  componentInjector: Injector;
-
-  componentContent = [
-    // eslint-disable-next-line no-undef
-    [document.createTextNode('Ng-content : Inspired by ')],
-    // eslint-disable-next-line no-undef
-    [document.createTextNode('https://angular.io/api/common/NgComponentOutlet')],
-  ];
-
-  // eslint-disable-next-line no-useless-constructor
-  constructor(private readonly injector: Injector) {}
-
-  ngOnInit(): void {
-    console.log({ color: this.color, name: this.name });
-
-    this.componentInjector = Injector.create({
-      providers: [
-        { provide: WITHOUT_SELECTOR_DATA, useValue: { color: this.color, name: this.name } },
-      ],
-      parent: this.injector,
-    });
-  }
-}
-
-// Live changing of args by controls does not work at the moment. When changing args storybook does not fully
-// reload and therefore does not take into account the change of provider.
-export const WithCustomNgComponentOutletWrapper: Story = (args) => ({
-  props: args,
-});
-WithCustomNgComponentOutletWrapper.argTypes = {
-  name: { control: 'text' },
-  color: { control: 'color' },
-};
-WithCustomNgComponentOutletWrapper.args = { name: 'Dixie Normous', color: 'green' };
-WithCustomNgComponentOutletWrapper.decorators = [
-  moduleMetadata({
-    declarations: [WithoutSelectorWrapperComponent],
-  }),
-  componentWrapperDecorator(WithoutSelectorWrapperComponent, (args) => ({
-    name: args.name,
-    color: args.color,
-    componentOutlet: WithoutSelectorComponent,
-  })),
-];
